Add options to run test selectively

+++ cmdline_tests/run_all.sh
@@ -7,6 +7,81 @@ L2_IP="10.10.1.101"
 ME=jintackl
 TEST_USER="root"
 
+LOCAL=0
+IDX_OFFSET=3
+
+TESTS="mysql netperf apache memcached"
+TEST_LIST=( $TESTS )
+
+__i=0
+for TEST in ${TESTS[@]}; do
+	TEST_ARRAY[$__i]=0
+	__i=$(($__i+1))
+done
+
+show_tests() {
+	i=0
+	echo [$i] "==== Start Test ====="
+
+	i=$(($i+1))
+	echo [$i] "All"
+
+	i=$(($i+1))
+	if [[ $LOCAL == 1 ]]; then
+		echo -n "*"
+	fi
+	echo [$i] "local tests (hackbench and kernbench)"
+
+	for TEST in ${TEST_LIST[@]}; do
+		i=$(($i+1))
+		idx=$(($i-$IDX_OFFSET))
+		if [[ ${TEST_ARRAY[$idx]} == 1 ]]; then
+			echo -n "*" 
+		fi
+		echo [$i] $TEST
+	done
+
+	echo -n "Type test number: "
+	read number
+
+	if [[ $number == 0 ]]; then
+		echo "Begin test"
+		break;
+	elif [[ $number == 1 ]]; then
+		__i=0
+		for TEST in ${TESTS[@]}; do
+			TEST_ARRAY[$__i]=1
+			__i=$(($__i+1))
+		done
+		LOCAL=1
+	elif [[ $number == 2 ]]; then
+		LOCAL=1
+	elif [[ $number -lt 7  ]]; then
+		idx=$(($number-$IDX_OFFSET))
+		TEST_ARRAY[$idx]=1
+	else
+		echo "Wrong test number"
+	fi
+	echo ""
+}
+
+while :
+do
+	show_tests
+done
+
+if [[ $LOCAL == 1 ]]; then
+	echo "Test Local"
+fi
+__i=0
+for TEST in ${TESTS[@]}; do
+	if [[ ${TEST_ARRAY[$__i]} == 1 ]]; then
+		echo "Test "${TEST_LIST[$__i]}
+	fi
+	__i=$(($__i+1))
+done
+
+exit
 echo "TEST LEVEL: $TEST_LEVEL"
 if [ $TEST_LEVEL == "L2" ] ; then
 	TARGET_IP=$L2_IP
@@ -49,16 +124,20 @@ if [ $TEST_LEVEL != "L0" ] ; then
 fi
 
 # Run local tests
-ssh $TEST_USER@$TARGET_IP "pushd ${LOCAL_PATH};rm *.txt"
-ssh $TEST_USER@$TARGET_IP "pushd ${LOCAL_PATH};sudo ./run_all.sh 0 3 0 10"
+if [[ $LOCAL == 1 ]]; then
+	ssh $TEST_USER@$TARGET_IP "pushd ${LOCAL_PATH};rm *.txt"
+	ssh $TEST_USER@$TARGET_IP "pushd ${LOCAL_PATH};sudo ./run_all.sh 0 3 0 10"
+fi
 
 # Prepare tests
 __i=0
 for TEST in ${TESTS[@]}; do
-	sudo ./${TEST}_install.sh
-	ssh $TEST_USER@$TARGET_IP "sudo ${CMD_PATH}/${TESTS[$__i]}_install.sh"
-	ssh $TEST_USER@$TARGET_IP "sudo service ${SERVICES[$__i]} stop"
-	ssh $TEST_USER@$TARGET_IP "sudo service ${SERVICES[$__i]} start"
+	if [[ ${TEST_ARRAY[$__i]} == 1 ]]; then
+		sudo ./${TEST}_install.sh
+		ssh $TEST_USER@$TARGET_IP "sudo ${CMD_PATH}/${TESTS[$__i]}_install.sh"
+		ssh $TEST_USER@$TARGET_IP "sudo service ${SERVICES[$__i]} stop"
+		ssh $TEST_USER@$TARGET_IP "sudo service ${SERVICES[$__i]} start"
+	fi
 	__i=$(($__i+1))
 done
 
@@ -67,15 +146,20 @@ ssh $TEST_USER@$TARGET_IP "sudo sed -i 's/^-l/#-l/' /etc/memcached.conf"
 ssh $TEST_USER@$TARGET_IP "sudo sed -i 's/^bind/#bind/' /etc/mysql/my.cnf"
 
 # Run mysql
-ssh $TEST_USER@$TARGET_IP "pushd ${CMD_PATH};sudo ./mysql.sh cleanup"
-ssh $TEST_USER@$TARGET_IP "pushd ${CMD_PATH};sudo ./mysql.sh prep"
-sudo ./mysql.sh run $TARGET_IP
-ssh $TEST_USER@$TARGET_IP "pushd ${CMD_PATH};sudo ./mysql.sh cleanup"
+__i=0
+if [[ ${TEST_ARRAY[$__i]} == 1 ]]; then
+	ssh $TEST_USER@$TARGET_IP "pushd ${CMD_PATH};sudo ./mysql.sh cleanup"
+	ssh $TEST_USER@$TARGET_IP "pushd ${CMD_PATH};sudo ./mysql.sh prep"
+	sudo ./mysql.sh run $TARGET_IP
+	ssh $TEST_USER@$TARGET_IP "pushd ${CMD_PATH};sudo ./mysql.sh cleanup"
+fi
 
 # Run tests except mysql
 __i=1
 for TEST in ${TESTS[@]}; do
-	./$TEST.sh $TARGET_IP
+	if [[ ${TEST_ARRAY[$__i]} == 1 ]]; then
+		./$TEST.sh $TARGET_IP
+	fi
 	__i=$(($__i+1))
 done
